requirements.txt
openpyxl
jinja2
pandas
python-docx
python-dotenv